Add: split as AccountMenu component
